requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0

# LLM
openai>=1.12.0
azure-identity>=1.15.0

# Memory
mem0ai>=0.1.0
qdrant-client>=1.7.0

# Microsoft
msal>=1.26.0
httpx>=0.26.0

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0

# Document parsing
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
pypdf>=4.0.0

# Scheduler
apscheduler>=3.10.0

# File uploads
python-multipart>=0.0.6
